refactor: move license-key-input from shared components to licensing module
