+++ app/src/modules/licensing/components/license-key-input.vue
